Duplicate `Session` for client and gateway
